Add consumed power calculation
When the grid is offline calculate the consumed power and write to
backend as monotonic increasing value. So in the end you can see the
graphic like this:

      |
100W  |        #
      |       ##
 50W  |    #####
      |  #######
      |#########
      ---------- 1H
          30m

As you can see - 100W of energy was consumed in 1 hour.

This type of info is collected if calc_consumed is enabled in
configuration.yaml file.

Use grid_status variable instead of reading of registers.

Fix aligning.